requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asn1crypto==1.5.1
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
croniter==6.0.0
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
flatbuffers==25.2.10
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.25.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.11.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.7
langchain-openai==0.2.0
langchain-text-splitters==0.3.8
mako==1.3.10
markupsafe==3.0.2
mmh3==5.2.0
more-itertools==10.8.0
multidict==6.6.4
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.3
oauthlib==3.3.1
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==24.2
pandas==2.3.1
patchy==2.10.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygithub==2.6.1
pyjwt==2.10.1
pynacl==1.6.0
pyopenssl==25.1.0
pypika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-sqlalchemy==1.7.6
sortedcontainers==2.4.0
sqlalchemy==2.0.43
sqlglot==27.4.1
sqllineage==1.5.4
sqlparse==0.5.3
starlette==0.47.1
tiktoken==0.12.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
